+++ packages/expo-file-system/ios/EXFileSystem/EXSessionTasks/EXSessionUploadTaskDelegate.m
@@ -10,17 +10,6 @@
 
 @implementation EXSessionUploadTaskDelegate
 
-- (instancetype)initWithSessionRegister:(id<EXSessionRegister>)sessionRegister
-                                resolve:(UMPromiseResolveBlock)resolve
-                                 reject:(UMPromiseRejectBlock)reject
-{
-  if (self = [super initWithSessionRegister:sessionRegister resolve:resolve reject:reject]) {
-    _responseData = [[NSMutableData alloc] init];
-  }
-  
-  return self;
-}
-
 - (void)URLSession:(NSURLSession *)session dataTask:(NSURLSessionDataTask *)dataTask didReceiveData:(NSData *)data
 {
   if (!data.length) {
@@ -33,18 +22,15 @@
 {
   if (error) {
     self.reject(@"ERR_FILE_SYSTEM_UNABLE_TO_UPLOAD_FILE.",
-          [NSString stringWithFormat:@"Unable to upload the file. '%@'", error.description],
+          [NSString stringWithFormat:@"Unable to upload the file: '%@'", error.description],
           error);
-    [self.sessionRegister unregister:session];
     return;
   }
   
   NSURLSessionUploadTask *uploadTask = (NSURLSessionUploadTask *)task;
-  NSMutableDictionary *result = [self parseServerResponse:uploadTask.response];
+  NSMutableDictionary *result = [[EXSessionTaskDelegate parseServerResponse:uploadTask.response] mutableCopy];
   result[@"body"] = [[NSString alloc] initWithData:_responseData encoding:NSUTF8StringEncoding];
   self.resolve(result);
-  
-  [self.sessionRegister unregister:session];
 }
 
 @end